adapted to moved PCMInstance
