+++ bundles/org.palladiosimulator.dependability.reliability.uncertainty.solver/src/org/palladiosimulator/dependability/reliability/uncertainty/solver/api/UncertaintyBasedReliabilityPredictionConfig.java
@@ -16,7 +16,7 @@ import org.palladiosimulator.dependability.reliability.uncertainty.UncertaintyPa
 import org.palladiosimulator.dependability.reliability.uncertainty.UncertaintyRepository;
 import org.palladiosimulator.dependability.reliability.uncertainty.solver.jobs.PCMInstanceBuilderJob;
 import org.palladiosimulator.dependability.reliability.uncertainty.solver.markov.StateSpaceExplorationStrategy;
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
 
 import de.uka.ipd.sdq.workflow.jobs.JobFailedException;
