+++ bundles/org.palladiosimulator.dependability.reliability.uncertainty/src/org/palladiosimulator/dependability/reliability/uncertainty/precondition/UncertaintyModelEqualityChecker.java
@@ -8,7 +8,7 @@ import org.palladiosimulator.envdyn.environment.staticmodel.GroundProbabilisticN
 import org.palladiosimulator.envdyn.environment.staticmodel.GroundRandomVariable;
 import org.palladiosimulator.envdyn.environment.staticmodel.LocalProbabilisticNetwork;
 import org.palladiosimulator.envdyn.environment.templatevariable.DependenceRelation;
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 
 import com.google.common.collect.Lists;
