+++ tests/org.palladiosimulator.dependability.reliability.uncertainty.solver.tests/test/org/palladiosimulator/dependability/reliability/uncertainty/solver/tests/ReliabilityPredictionTestDefinition.java
@@ -31,7 +31,7 @@ import org.palladiosimulator.dependability.reliability.uncertainty.solver.tests.
 import org.palladiosimulator.experimentautomation.application.jobs.PrepareBlackboardJob;
 import org.palladiosimulator.reliability.solver.pcm2markov.MarkovTransformationResult;
 import org.palladiosimulator.reliability.solver.pcm2markov.Pcm2MarkovStrategy;
-import org.palladiosimulator.solver.models.PCMInstance;
+import org.palladiosimulator.solver.core.models.PCMInstance;
 import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
 
 import com.google.common.collect.Lists;
